refactor: Change the path of 'Collider' related files to 'Geometry'
