+++ Includes/Core/Geometry/RigidBodyCollider2.hpp
@@ -11,7 +11,7 @@
 #ifndef CUBBYFLOW_RIGID_BODY_COLLIDER2_HPP
 #define CUBBYFLOW_RIGID_BODY_COLLIDER2_HPP
 
-#include <Core/Collider/Collider2.hpp>
+#include <Core/Geometry/Collider2.hpp>
 
 namespace CubbyFlow
 {
