+++ Includes/Core/Geometry/RigidBodyCollider3.hpp
@@ -11,7 +11,7 @@
 #ifndef CUBBYFLOW_RIGID_BODY_COLLIDER3_HPP
 #define CUBBYFLOW_RIGID_BODY_COLLIDER3_HPP
 
-#include <Core/Collider/Collider3.hpp>
+#include <Core/Geometry/Collider3.hpp>
 
 namespace CubbyFlow
 {
