+++ Includes/Core/Solver/Grid/GridFluidSolver2.hpp
@@ -12,8 +12,8 @@
 #define CUBBYFLOW_GRID_FLUID_SOLVER2_HPP
 
 #include <Core/Animation/PhysicsAnimation.hpp>
-#include <Core/Collider/Collider2.hpp>
 #include <Core/Emitter/GridEmitter2.hpp>
+#include <Core/Geometry/Collider2.hpp>
 #include <Core/Grid/GridSystemData2.hpp>
 #include <Core/Solver/Advection/AdvectionSolver2.hpp>
 #include <Core/Solver/Grid/GridDiffusionSolver2.hpp>
