+++ Includes/Core/Solver/Particle/ParticleSystemSolver2.hpp
@@ -12,9 +12,9 @@
 #define CUBBYFLOW_PARTICLE_SYSTEM_SOLVER2_HPP
 
 #include <Core/Animation/PhysicsAnimation.hpp>
-#include <Core/Collider/Collider2.hpp>
 #include <Core/Emitter/ParticleEmitter2.hpp>
 #include <Core/Field/VectorField2.hpp>
+#include <Core/Geometry/Collider2.hpp>
 #include <Core/Particle/ParticleSystemData2.hpp>
 #include <Core/Utils/Constants.hpp>
 #include <Core/Vector/Vector2.hpp>
