+++ Includes/Core/Solver/Particle/ParticleSystemSolver3.hpp
@@ -12,9 +12,9 @@
 #define CUBBYFLOW_PARTICLE_SYSTEM_SOLVER3_HPP
 
 #include <Core/Animation/PhysicsAnimation.hpp>
-#include <Core/Collider/Collider3.hpp>
 #include <Core/Emitter/ParticleEmitter3.hpp>
 #include <Core/Field/VectorField3.hpp>
+#include <Core/Geometry/Collider3.hpp>
 #include <Core/Particle/ParticleSystemData3.hpp>
 #include <Core/Utils/Constants.hpp>
 #include <Core/Vector/Vector3.hpp>
