+++ Sources/Core/Geometry/ColliderSet3.cpp
@@ -8,7 +8,7 @@
 // personal capacity and are not conveying any rights to any intellectual
 // property of any third parties.
 
-#include <Core/Collider/ColliderSet3.hpp>
+#include <Core/Geometry/ColliderSet3.hpp>
 #include <Core/Surface/SurfaceSet3.hpp>
 
 namespace CubbyFlow
