+++ Tests/ManualTests/FLIPSolver2Tests.cpp
@@ -2,9 +2,9 @@
 
 #include <ManualTests.hpp>
 
-#include <Core/Collider/RigidBodyCollider2.hpp>
 #include <Core/Emitter/VolumeParticleEmitter2.hpp>
 #include <Core/Geometry/Box2.hpp>
+#include <Core/Geometry/RigidBodyCollider2.hpp>
 #include <Core/Geometry/Sphere2.hpp>
 #include <Core/Solver/Hybrid/FLIP/FLIPSolver2.hpp>
