+++ Tests/ManualTests/LevelSetLiquidSolver2Tests.cpp
@@ -1,9 +1,9 @@
 #include "pch.hpp"
 
-#include <Core/Collider/RigidBodyCollider2.hpp>
 #include <Core/Emitter/VolumeGridEmitter2.hpp>
 #include <Core/Geometry/Box2.hpp>
 #include <Core/Geometry/Plane2.hpp>
+#include <Core/Geometry/RigidBodyCollider2.hpp>
 #include <Core/Geometry/Sphere2.hpp>
 #include <Core/LevelSet/LevelSetUtils.hpp>
 #include <Core/Solver/Grid/GridFractionalSinglePhasePressureSolver2.hpp>
