+++ Tests/ManualTests/ParticleSystemSolver2Tests.cpp
@@ -2,10 +2,10 @@
 
 #include <ManualTests.hpp>
 
-#include <Core/Collider/RigidBodyCollider2.hpp>
 #include <Core/Emitter/PointParticleEmitter2.hpp>
 #include <Core/Field/ConstantVectorField2.hpp>
 #include <Core/Geometry/Plane2.hpp>
+#include <Core/Geometry/RigidBodyCollider2.hpp>
 #include <Core/Solver/Particle/ParticleSystemSolver2.hpp>
 
 using namespace CubbyFlow;
