+++ Tests/ManualTests/SPHSolver2Tests.cpp
@@ -2,10 +2,10 @@
 
 #include <ManualTests.hpp>
 
-#include <Core/Collider/RigidBodyCollider2.hpp>
 #include <Core/Emitter/VolumeParticleEmitter2.hpp>
 #include <Core/Geometry/Box2.hpp>
 #include <Core/Geometry/Plane2.hpp>
+#include <Core/Geometry/RigidBodyCollider2.hpp>
 #include <Core/Geometry/Sphere2.hpp>
 #include <Core/Solver/Particle/SPH/SPHSolver2.hpp>
 #include <Core/Surface/ImplicitSurfaceSet2.hpp>
