+++ Tests/UnitTests/ColliderSet3Tests.cpp
@@ -1,8 +1,8 @@
 #include "pch.hpp"
 
-#include <Core/Collider/ColliderSet3.hpp>
-#include <Core/Collider/RigidBodyCollider3.hpp>
 #include <Core/Geometry/Box3.hpp>
+#include <Core/Geometry/ColliderSet3.hpp>
+#include <Core/Geometry/RigidBodyCollider3.hpp>
 
 using namespace CubbyFlow;
