+++ Tests/UnitTests/RigidbodyCollider2Tests.cpp
@@ -1,7 +1,7 @@
 #include "pch.hpp"
 
-#include <Core/Collider/RigidBodyCollider2.hpp>
 #include <Core/Geometry/Plane2.hpp>
+#include <Core/Geometry/RigidBodyCollider2.hpp>
 #include <Core/Surface/ImplicitSurfaceSet2.hpp>
 
 using namespace CubbyFlow;
