+++ Tests/UnitTests/RigidbodyCollider3Tests.cpp
@@ -1,7 +1,7 @@
 #include "pch.hpp"
 
-#include <Core/Collider/RigidBodyCollider3.hpp>
 #include <Core/Geometry/Plane3.hpp>
+#include <Core/Geometry/RigidBodyCollider3.hpp>
 #include <Core/Surface/ImplicitSurfaceSet3.hpp>
 
 using namespace CubbyFlow;
